requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.8.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
confusable-homoglyphs==3.2.0
Django==4.1.1
django-environ==0.9.0
django-map-widgets==0.4.0
django-registration==3.3
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.3
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.12
flake8==5.0.4
